[node] plow ahead with `initCap` refactor;
